Adds Community Nights Seeder and Images
Adds a seeder to populate the database with initial Community Night data, and associated images.

Updates the community nights index and detail views to correctly display the images, using the asset helper to reference the storage path.

Modifies the pagination on the index page to display only 1 item per page with 1 item on each side.